requirements.txt
python-telegram-bot==21.0
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.3
